had to rename the old-structure dir in order to make the test on the old code green
