+++ matchms/old/helper_functions.py
