+++ matchms/old/similarity_measure.py
